Replaced Int with Long for some variables (#178)

+++ bigbone/src/main/kotlin/social/bigbone/api/entity/InstanceStats.kt
@@ -11,17 +11,17 @@ data class InstanceStats(
      * Total users on this instance.
      */
     @SerializedName("user_count")
-    val userCount: Int = 0,
+    val userCount: Long = 0,
 
     /**
      * Total statuses on this instance.
      */
     @SerializedName("status_count")
-    val statusCount: Int = 0,
+    val statusCount: Long = 0,
 
     /**
      * Total domains discovered by this instance.
      */
     @SerializedName("domain_count")
-    val domainCount: Int = 0
+    val domainCount: Long = 0
 )
